refactor: defer import of cli entrypoints
Repo-related CLI files indirectly pull in the `git` package, that probes
for the git binary *at import time*, dying loudly in its own way that is
a bit scary for end users of `ruyi`.

To improve the UX, we're going to probe the prerequisite binaries
ourselves, and only after ensuring everything's okay are we going to
actually import the CLI entrypoints. This might be the simplest way to
ensure as few packages as possible are imported before check time.