fix: fallback to hard-coded well-known ca-certificates paths if probed paths are not usable
In some cases we apparently dlopen the bundled libcrypto.so.3 (because
the system is not having a libcrypto.so installed), so the paths could
be unreliable and absent on the running system even if they're probed.

Fall back to a list of well-known paths in this case.

Fixes: #103
(cherry picked from commit d6970589c6e6ee31ba301aa4874ed7ca2c228bee)