build: bump self version to 0.7.1
With the desire to keep RuyiSDK's press-oriented version number and
Ruyi's semver aligned, and the fact that the release is intended to fix
a user story for "RuyiSDK 0.7", let's name the release 0.7.1 (also this
version number is going to provoke some certain image in riscv devs).
This is also intended to be a test for future real patch releases.