feat(provision): only wrap dd & fastboot invocations with sudo if bare calls failed
This is going to help power users with exotic setup that:

* HAVE the privileges to necessary devices and/or block device files,
  but
* CANNOT do `sudo`.

For some context, the reporter of #120 seemingly is trying Ruyi on NixOS,
so they have to run it in a sandboxed wrapper that emulates the FHS
filesystem layout, and that's why the shell is NO_NEW_PRIVS but with all
privileges already granted from outside. While NixOS is not on the official
supported platform list of RuyiSDK, it is beneficial and safer to require
less privileges in general, so let's do this. The approach implemented
is also suggested by them.

Fixes: #120