ci: remove LGPL from the license allowlist
According to private conversation, the driving force behind the
project wants to make the project "commercial-friendly", that can be
viewed as minimizing legal obligations for potential future downstream
forks that want to go closed-source. Importantly, a request for
introducing Qt was turned down; as open-source Qt is LGPL-licensed, it
can only be deduced that any other LGPL usage is likely also unwelcomed.

Now that pyxdg is gone from the list of deps, let's remove LGPL from the
license allowlist, for one less thing to worry about.