fix(cli): ensure line-wrapping for RuyiStyledMarkdown
The log stdout console was made non-wrapping by default to fix #181, but
it had the unfortunate side-effect of cropping list items and
blockquotes that are text contents and really need to be line-wrapped.

Avoid the side effect by overriding RuyiStyledMarkdown's render options
so that line-wrapping is enabled for Markdown rendering alone, and make
NonWrappingCodeBlock undo the line-wrapping options. This way Markdown
rendering stays the same as before #187, but logs and code blocks (and
only these) can be properly non-wrapped.