refactor(ruyipkg): rename some misleading "manifest" names
A "manifest" in ruyipkg sense is a complete description for a particular
version of a package, and what's generated by the `ruyi admin manifest`
command is in fact a list of `[[distfile]]` entries.

The command's name might be influenced by Gentoo Portage (which is an
inspiration for Ruyi), where "manifest"'s are indeed checksums, but
renaming the command is a separate thing to do.