feat(ruyipkg): add basic `ruyi admin format-manifest`
This is like similar tools such as `cargo fmt` or `gofmt` that define a
canonical style for relevant files, but in addition to formatting, this
command also takes care of auto-migrating JSON / v0 package manifests to
v1 TOML ones.

This is implemented for `blob` and `provisionable` kinds for now; other
kinds are to be supported in following commits.