feat(venv): [breaking change] allow multiple non-conflicting toolchains
Revise the venv cache format to allow configuring each supported target
tuple separately. Potential conflicting "provides" are resolved mainly
on a "first-one-wins" basis.

It is not recommended to mix-and-match `ruyi` and venvs created with
previous versions of `ruyi`, but hopefully compatibility is preserved.
Auto-upgrading of venv cached config format is left out for now.

Fixes: #199