feat(pluginhost): add an unsandboxed plugin host implementation
This unexpected change is requested by the Ruyi project owner, as a
direct consequence of a recently-clarified requirement of implementing
plugin functionality with only Python 3, but not any dialect of it.

While technically speaking obviously this is much more insecure, it
is thought to be still within the current threat model of RuyiSDK, as
no "outsiders" are involved in plugin creation yet, and attacks from
"insiders" are not going to be thwarted by code-level sandboxing alone.