feat(cli): rename `ruyi admin manifest` to `ruyi admin checksum`
Formerly we mimicked Gentoo's `ebuild /path/to/ebuild/file manifest`,
but as we used the term "manifest" for the actual definition of specific
package versions, the command name becomes misleading as a result -- the
output may be thought to comprise a complete manifest that's formattable
with `ruyi admin format-manifest` but in fact it's not.

The easiest way to avoid such user misunderstanding is to rename the
`manifest` command. As it's meant for RuyiSDK packagers, and currently
all packaging work is done by "insiders", we can break compatibility
here without regressing real UX.

Fixes: #217