feat(telemetry): default to mode "on"
It turns out the RuyiSDK project has hard KPI on user metrics, so we
have no option but to collect. We don't collect PII though, and we
expect to maintain a high standard for our analytics.

No servers have been configured yet, and this will happen through a
repo update.