refactor(cli): namespace all python-argcomplete helpers
The code is taken from argcomplete's common infra, but now it is only
supposed to be used for ruyi, so unshare the functions by modifying the
names' prefix.

While at it, document the likely source of the script, but I am not
entirely sure, as the original author did not provide detailed commit
history for others to reproduce the adaptations they did.

Signed-off-by: WANG Xuerui <wangxuerui@iscas.ac.cn>