feat(provision): add alias `ruyi device flash` for `ruyi device provision`
Many users reported difficulties with the word "provision", and
suggested "flash" as an alternative. While what we do is expected to
be more than just "flashing" at least in the roadmap, it is the case for
now, and "flash" is easier for non-English speakers. So just add an
alias to reduce typing and confusion for those.

Link: https://ruyisdk.cn/t/topic/506
Signed-off-by: WANG Xuerui <wangxuerui@iscas.ac.cn>