refactor(ruyipkg): allow to query repo without ensuring its existence
At CLI auto-completion invocation time, the user may not have the repo
initialized, because the "real" invocation has not been made yet; but at
this stage some auto-completers may already want to query repo data,
causing the repo to be synced at CLI auto-completion time. This is bad
for UX because it's very noticeable lag, contaminates stdout, and
surprising behavior in general.

To fix this, allow various repo operations to skip ensuring the repo's
existence, and return empty data when the repo is non-existent locally.

Signed-off-by: WANG Xuerui <wangxuerui@iscas.ac.cn>