fix(ruyipkg): only consider repo path to be customized if different from the default
While the #372 description is a bit unclear, there actually is
inconsistency between the computation of local repo path and the
"customization" detection. The former only looks at truthiness of
override_repo_dir, making an empty string equivalent to an unset value,
while the latter is a strict None check treating an empty string
differently. Make them consistent by moving the logic into GlobalConfig.

While at it, treat a "set but same as default" repo path setting as "not
customized", because the path is considered to be managed by us anyway,
and that the user may not be aware of the difference between "unsetting
a value" and "setting the default".

See: #372
Signed-off-by: WANG Xuerui <wangxuerui@iscas.ac.cn>