feat(telemetry): implement `ruyi telemetry --cron-upload` for guaranteed weekly pings
Project decision makers requested this data to be reported even if the
user may not invoke `ruyi` on the scheduled weekday, so a cron job must
be used in this case. While this cron job must be set up per user
(because data is not recorded in a system-wide well-known location) thus
expected to be of limited adoption, here's the necessary machinery to
make it possible at all.

Signed-off-by: WANG Xuerui <wangxuerui@iscas.ac.cn>